Add files via upload
I made a simple application that can read the data and display it in the form of a table. Due to limited time constraint and my lack of knowledge on the subject, I did it manually instead of loading it directly from the JSON file. The data was copied from the JSON file and stored into the HTML file and then processed to make a table.